combined approach to local and remote tests
