--- tests/local/pom.xml
@@ -1,116 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2016 Google Inc. All Rights Reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.google.cloud.runtimes.tests</groupId>
-        <artifactId>tests-parent</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-    <groupId>com.google.cloud.runtimes.tests.local</groupId>
-    <artifactId>local-parent</artifactId>
-    <name>Jetty-Runtime :: Local Tests :: Local Parent</name>
-    <packaging>pom</packaging>
-
-    <properties>
-        <checkstyle.skip>true</checkstyle.skip>
-    </properties>
-
-    <modules>
-        <module>test-war-hello</module>
-    </modules>
-
-    <build>
-        <plugins>
-            <!--
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-failsafe-plugin</artifactId>
-                <version>2.19.1</version>
-                <dependencies>
-                    <dependency>
-                        <groupId>org.apache.maven.surefire</groupId>
-                        <artifactId>surefire-junit47</artifactId>
-                        <version>2.19.1</version>
-                    </dependency>
-                </dependencies>
-                <configuration>
-                    <systemPropertyVariables>
-                        <app.deploy.project>${app.deploy.project}</app.deploy.project>
-                        <app.deploy.version>${maven.build.timestamp}</app.deploy.version>
-                    </systemPropertyVariables>
-                </configuration>
-            </plugin>
-            -->
-            <!-- This can be re-enabled when app-maven-plugin issue is fixed
-            https://github.com/GoogleCloudPlatform/app-maven-plugin/issues/85
-            <plugin>
-              <groupId>com.google.cloud.tools</groupId>
-              <artifactId>appengine-maven-plugin</artifactId>
-              <version>0.1.2</version>
-              <configuration>
-                <dockerDirectory>${project.build.directory}/docker</dockerDirectory>
-              </configuration>
-              <executions>
-                <execution>
-                  <id>deploy-integration-test</id>
-                  <phase>integration-test</phase>
-                  <goals>
-                    <goal>deploy</goal>
-                  </goals>
-                  <configuration>
-                    <deploy.project>${app.deploy.project}</deploy.project>
-                    <deploy.version>${maven.build.timestamp}</deploy.version>
-                    <deploy.promote>false</deploy.promote>
-                  </configuration>
-                </execution>
-              </executions>
-            </plugin>
-            -->
-            <!--
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-antrun-plugin</artifactId>
-                <version>1.8</version>
-                <executions>
-                    <execution>
-                        <id>cleanup-integration-test</id>
-                        <phase>post-integration-test</phase>
-                        <goals>
-                            <goal>run</goal>
-                        </goals>
-                        <configuration>
-                            <target>
-                                <exec executable="gcloud">
-                                    <arg value="-quiet"/>
-                                    <arg value="app"/>
-                                    <arg value="versions"/>
-                                    <arg value="delete"/>
-                                    <arg value="${maven.build.timestamp}"/>
-                                </exec>
-                            </target>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            -->
-        </plugins>
-    </build>
-
-</project>
