+++ tests/pom.xml
@@ -1,70 +1,100 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2016 Google Inc. All Rights Reserved.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.google.cloud.runtimes</groupId>
-        <artifactId>jetty-parent</artifactId>
-        <version>0.1.0-SNAPSHOT</version>
-    </parent>
-    <groupId>com.google.cloud.runtimes.tests</groupId>
-    <artifactId>tests-parent</artifactId>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>gcloud-testing-core</module>
-    </modules>
-    <build>
+<!--  ~ Copyright 2016 Google Inc. All Rights Reserved.  ~  ~ Licensed under the Apache License, Version 2.0 (the "License");  ~ you may not use this file except in compliance with the License.  ~ You may obtain a copy of the License at  ~  ~       http://www.apache.org/licenses/LICENSE-2.0  ~  ~ Unless required by applicable law or agreed to in writing, software  ~ distributed under the License is distributed on an "AS IS" BASIS,  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  ~ See the License for the specific language governing permissions and  ~ limitations under the License.  -->
+<project xmlns="http://maven.apache.org/POM/4.0.0"         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>com.google.cloud.runtimes</groupId>
+    <artifactId>jetty-parent</artifactId>
+    <version>0.1.0-SNAPSHOT</version>
+  </parent>
+  <groupId>com.google.cloud.runtimes.tests</groupId>
+  <artifactId>tests-parent</artifactId>
+  <packaging>pom</packaging>
+  <properties>
+    <checkstyle.skip>true</checkstyle.skip>
+  </properties>
+  <modules>
+    <module>gcloud-testing-core</module>
+    <module>test-war-hello</module>
+  </modules>
+  <build>
+    <pluginManagement>
+      <plugins>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-failsafe-plugin</artifactId>
+          <version>2.19.1</version>
+          <dependencies>
+            <dependency>
+              <groupId>org.apache.maven.surefire</groupId>
+              <artifactId>surefire-junit47</artifactId>
+              <version>2.19.1</version>
+            </dependency>
+          </dependencies>
+          <configuration>
+            <skipTests>true</skipTests>
+          </configuration>
+        </plugin>
+      </plugins>
+    </pluginManagement>
     <resources>
-        <resource>
-            <directory>src/main/docker</directory>
-            <filtering>true</filtering>
-            <targetPath>${project.build.directory}/docker</targetPath>
-        </resource>
-        <resource>
-            <directory>src/main/appengine</directory>
-            <filtering>true</filtering>
-            <targetPath>${project.build.directory}/appengine-staging</targetPath>
-        </resource>
-        <resource>
-            <directory>src/main/resources</directory>
-            <filtering>true</filtering>
-            <targetPath>${project.build.directory}/${project.artifactId}-${project.version}/WEB-INF/classes</targetPath>
-        </resource>
+      <resource>
+        <directory>src/main/docker</directory>
+        <filtering>true</filtering>
+        <targetPath>${project.build.directory}/docker</targetPath>
+      </resource>
+      <resource>
+        <directory>src/main/appengine</directory>
+        <filtering>true</filtering>
+        <targetPath>${project.build.directory}/appengine-staging</targetPath>
+      </resource>
+      <resource>
+        <directory>src/main/resources</directory>
+        <filtering>true</filtering>
+        <targetPath>${project.build.directory}/${project.artifactId}-${project.version}/WEB-INF/classes</targetPath>
+      </resource>
     </resources>
-    </build>
-    <!--
-    - tests are not run by default, choose the style of testings by activating the appropraite
-    - testing profile
-    -->
-    <profiles>
-        <profile>
-            <id>local</id>
-            <modules>
-                <module>local</module>
-            </modules>
-        </profile>
-        <profile>
-            <id>remote</id>
-            <modules>
-                <module>remote</module>
-            </modules>
-        </profile>
-    </profiles>
+  </build>
+  <profiles>
+    <profile>
+      <id>local</id>
+      <build>
+        <plugins>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-failsafe-plugin</artifactId>
+            <version>2.19.1</version>
+            <configuration>
+              <skipTests>false</skipTests>
+              <includes>
+                <include>%regex[.*LocalITCase.*]</include>
+              </includes>
+            </configuration>
+          </plugin>
+        </plugins>
+      </build>
+    </profile>
+    <profile>
+      <id>remote</id>
+      <properties>
+        <maven.build.timestamp.format>yyyyMMdd-HHmm</maven.build.timestamp.format>
+        <gcloud-projectId-file>${project.build.directory}/gcloud-projectid.properties</gcloud-projectId-file>
+      </properties>
+      <build>
+        <plugins>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-failsafe-plugin</artifactId>
+            <version>2.19.1</version>
+            <configuration>
+              <skipTests>false</skipTests>
+              <includes>
+                <include>%regex[.*RemoteITCase.*]</include>
+              </includes>
+            </configuration>
+          </plugin>
+        </plugins>
+      </build>
+    </profile>
+  </profiles>
 </project>
