--- tests/remote/test-war-hello/pom.xml
@@ -1,78 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>com.google.cloud.runtimes.tests.remote</groupId>
-    <artifactId>remote-parent</artifactId>
-    <version>0.1.0-SNAPSHOT</version>
-  </parent>
-  <artifactId>test-war-hello</artifactId>
-  <packaging>war</packaging>
-  <name>Jetty-Runtime :: Tests :: Hello (Servlet 3.1) WebApp</name>
-  <dependencies>
-    <dependency>
-      <groupId>javax.servlet</groupId>
-      <artifactId>javax.servlet-api</artifactId>
-      <version>3.1.0</version>
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.google.cloud.runtimes.jetty.tests</groupId>
-      <artifactId>gcloud-testing-core</artifactId>
-      <version>${project.version}</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.5.1</version>
-        <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-antrun-plugin</artifactId>
-        <version>1.8</version>
-        <executions>
-          <execution>
-            <goals>
-              <goal>run</goal>
-            </goals>
-            <phase>package</phase>
-            <configuration>
-              <target>
-                <unjar src="${project.build.directory}/${project.build.finalName}.war"
-                       dest="${project.build.directory}/docker/root"/>
-              </target>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-      <plugin>
-        <groupId>com.google.cloud.tools</groupId>
-        <artifactId>appengine-maven-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>deploy-integration-test</id>
-            <phase>pre-integration-test</phase>
-            <goals>
-              <goal>deploy</goal>
-            </goals>
-            <configuration>
-              <deploy.project>${app.deploy.project}</deploy.project>
-              <deploy.version>${maven.build.timestamp}</deploy.version>
-              <deploy.promote>false</deploy.promote>
-              <dockerDirectory>${project.build.directory}/docker</dockerDirectory>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-</project>
