--- tests/remote/test-war-hello/src/main/appengine/app.yaml
@@ -1,3 +1,0 @@
-runtime: custom
-vm: true
-service: hello
