--- tests/remote/test-war-hello/src/main/docker/Dockerfile
@@ -1,2 +1,0 @@
-FROM ${jetty.test.image}
-ADD root /app
