--- tests/remote/test-war-hello/src/main/java/com/google/cloud/runtimes/jetty/tests/HelloServlet.java
@@ -1,20 +1,0 @@
-package com.google.cloud.runtimes.jetty.tests;
-
-import java.io.IOException;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@WebServlet(urlPatterns = {"/hello/*"})
-public class HelloServlet extends HttpServlet
-{
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
-    {
-        resp.setContentType("text/plain");
-        resp.getWriter().println("Hello from Servlet 3.1");
-    }
-}
