+++ tests/test-war-hello/src/main/docker/Dockerfile
