+++ tests/test-war-hello/src/test/java/com/google/cloud/runtimes/jetty/tests/HelloRemoteITCase.java
@@ -1,8 +1,5 @@
 package com.google.cloud.runtimes.jetty.tests;
 
-import static org.hamcrest.CoreMatchers.containsString;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
 
 import com.google.cloud.runtime.jetty.testing.AppDeployment;
 import com.google.cloud.runtime.jetty.testing.HttpUrlUtil;
@@ -11,11 +8,9 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.io.IOException;
-import java.net.HttpURLConnection;
 import java.util.concurrent.TimeUnit;
 
-public class HelloITCase
-{
+public class HelloRemoteITCase extends AbstractHelloTestCase {
     @BeforeClass
     public static void isServerUp()
     {
@@ -25,9 +20,7 @@ public class HelloITCase
     @Test
     public void testGet() throws IOException
     {
-        HttpURLConnection http = HttpUrlUtil.openTo(AppDeployment.SERVER_URI.resolve("/hello/"));
-        assertThat(http.getResponseCode(), is(200));
-        String responseBody = HttpUrlUtil.getResponseBody(http);
-        assertThat(responseBody, containsString("Hello from Servlet 3.1"));
+        assertTestGet(AppDeployment.SERVER_URI.resolve("/hello/"));
     }
+
 }
